feat: (worker) reorganise common code in cmd package
